% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{abbrevNames}
\alias{abbrevNames}
\title{Abbreviate a string to use as a biomod2 species name}
\usage{
abbrevNames(x)
}
\arguments{
\item{x}{A string}
}
\value{
A simplified string
}
\description{
A simple function used to abbreviate a species name, remove spaces and points 
to use it in biomod2
}
