% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{checkModAlgoRun}
\alias{checkModAlgoRun}
\title{Check which model algorithms where run}
\usage{
checkModAlgoRun(biomodModelOut)
}
\arguments{
\item{biomodModelOut}{A \link[biomod2]{BIOMOD.models.out-class} class object.}
}
\value{
A named logical vector specifying which models were run.
}
\description{
A simple function used to verify which models were actually run by \code{biomod2} 
for a given target species.
}
