% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{checkRasterStack}
\alias{checkRasterStack}
\title{Check raster stacks for Maxent}
\usage{
checkRasterStack(x)
}
\arguments{
\item{x}{A \code{RasterStack} object to check}
}
\value{
A new and checked \code{RasterStack}
}
\description{
Checks a \code{RasterStack} object to avoid problems during Maxent (Phillips) projection step
}
\note{
This is based on Damien Georges suggestion at: 
\url{https://r-forge.r-project.org/forum/message.php?msg_id=42888&group_id=302}
}
