% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{convertToGeoTIFF}
\alias{convertToGeoTIFF}
\title{Convert raster files from grd format to GeoTIFF}
\usage{
convertToGeoTIFF(inputFolder, outputFolder)
}
\arguments{
\item{inputFolder}{Input folder with .grd files}

\item{outputFolder}{Output folder where GeoTIFF raster 
files will be placed}
}
\description{
A function used to convert rasters in grd file format from a source 
folder to GeoTIFF format to an output folder
}
