% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{countDistinct}
\alias{countDistinct}
\title{Count distinct elements}
\usage{
countDistinct(x)
}
\arguments{
\item{x}{An input vector}
}
\value{
An integer with the number of unique elements in the input vector
}
\description{
A help function used to count the number of distinct elements in a
vector
}
