% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{evalMetricPlot}
\alias{evalMetricPlot}
\title{A plot for evaluation metrics}
\usage{
evalMetricPlot(biomodModelOut, evalMetric = "TSS", by = "algo",
  sort = TRUE, removeFull = FALSE)
}
\description{
Makes a plot using ggplot2 for one selected evaluation metric for each 
modelling algorithm
}
