% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{getEvalMetricQuantile}
\alias{getEvalMetricQuantile}
\title{Calculate the quantile of an evaluation metric}
\usage{
getEvalMetricQuantile(myBiomodModelOut, qt = 0.75, evalMetric = "TSS")
}
\arguments{
\item{myBiomodModelOut}{A \link[biomod2]{BIOMOD.models.out-class} class object.}

\item{qt}{The target quantile between 0 and 1.}

\item{evalMetric}{A character string defining the evaluation metric. 
Available options are: 'KAPPA', 'TSS', 'ROC', 'FAR', 'SR', 'ACCURACY', 'BIAS', 
'POD', 'CSI' and 'ETS'.}
}
\value{
The target quantile value.
}
\description{
An ancillary function to calculate a target quantile of an evaluation metric 
calculated by \code{biomod2} hold-out evaluation. This function can be used to 
calculate the threshold used in \link[biomod2]{BIOMOD_EnsembleModeling} function 
to set the parameter \code{eval.metric.quality.threshold}.
}
