% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{getUniqueRasterXYCoords}
\alias{getUniqueRasterXYCoords}
\title{Get distinct raster cell centroids from input presence records}
\usage{
getUniqueRasterXYCoords(rasterObj, pointObj, spatial = TRUE)
}
\arguments{
\item{rasterObj}{A raster stack or layer object}

\item{pointObj}{A set of points as a SpatialPoints* object}

\item{spatial}{Return as a spatial object? (default: TRUE)}
}
\value{
Either a \code{data.frame} with unique cell centroid coordinates for the input 
point dataset or a \code{SpatialPoints} object
}
\description{
This function is used to remove duplicates and NA's in the input points and 
return the coordinates for each unique cell centroid
}
