% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in function}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}
}
\value{
A logical vector containing the elements of x that are not in y
}
\description{
A simple not in function operator
}
