% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{intersectMask}
\alias{intersectMask}
\title{Intersect raster layers to avoid Maxent projection problems}
\usage{
intersectMask(x)
}
\arguments{
\item{x}{A RasterStack object}
}
\value{
A corrected RasterStack
}
\description{
A help function provided by \code{biomod2} creators to avoid projection 
errors in MAXENT.Phillips
}
