% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{powerFunPAnumberCalc}
\alias{powerFunPAnumberCalc}
\title{Calculate the number of pseudo-absences using a power function}
\usage{
powerFunPAnumberCalc(nPresences, a = 45, b = -0.55)
}
\arguments{
\item{nPresences}{Number of presence records}

\item{a}{Multiplication parameter (default: 45)}

\item{b}{Exponential parameter (default: -0.55)}
}
\value{
An integer defining the number of pseudo-absences
}
\description{
This function is intended to calculate the number of pseudo-absences based on a 
power function thus allowing to have a large number of PA's for small sample 
sizes and, it approximates the number presences for very large sample sizes. 
The function is defined  as follows: \eqn{n_{PA} = n_{P} \times a \times n_{P}^{-b}}, 
with \eqn{n_{PA}} equal to the number of pseudo-absences and \eqn{n_{P}} to the number of 
presences. The parameters \eqn{a} and \eqn{b} allow to adjust the amount of PA's.
}
