% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{twoStepBestModelSelection}
\alias{twoStepBestModelSelection}
\title{Performs a two-stage best model selection}
\usage{
twoStepBestModelSelection(biomodModelOut, evalMetric = "TSS",
  nrBestAlgos = 5, bestAlgoFun = stats::median, topFraction = 0.25,
  na.rm = TRUE)
}
\arguments{
\item{biomodModelOut}{A \link[biomod2]{BIOMOD.models.out-class} class object}

\item{evalMetric}{The evaluation metric to use (available options are: 
'KAPPA', TSS', 'ROC', 'FAR', 'SR', 'ACCURACY', 'BIAS', 'POD', 'CSI' and 'ETS')}

\item{nrBestAlgos}{Integer number defining how many top-performing 
algorithms (default: 5)}

\item{bestAlgoFun}{Function used to aggregate model scores and rank each 
algorithm (default: median)}

\item{topFraction}{Fraction of best-performing models for each selected modelling 
technique}
}
\value{
A character vector with selected models
}
\description{
The function is used to make a two-stage selection based, first selecting the best 
modelling algorithms and then selecting the top best-performing fraction of models 
within those techniques.
}
