% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{responsePlots}
\alias{responsePlots}
\title{Make response plots for multiple variables}
\usage{
responsePlots(biomodModelOut, Data, modelsToUse, showVars = "all",
  fixedVarMetric = "mean", addMarginalPlot = TRUE, varNames = NULL,
  marginPlotType = "histogram", plotStdErr = FALSE, plot = FALSE,
  save = TRUE, filePrefix = "ResponsePlot_", fileSuffix = "",
  outFolder = getwd(), width = NA, height = NA, dpi = 300)
}
\arguments{
\item{biomodModelOut}{A \link[biomod2]{BIOMOD.models.out} class object.}

\item{Data}{A \code{data.frame} or \code{RasterStack} containing the data used to make 
the response plots. They must have the same names as the ones used to calibrate the model.}

\item{dpi}{}
}
\value{
A list object containing two components:     
\itemize{
  \item DATA - inside this, a list with one element for each predictor variable with the data 
  used to make the response plots
  \item PLOTS - the \code{ggplot} objects. One for each variable.
}
}
\description{
A flexiblw function used to produce response plots using \code{ggplot2} package.
}
\details{
This function uses Adaptation of the Evaluation Strip method proposed by Elith et al.(2005) 
and implemented in the \link[biomod2]{response.plot2}. 
This function enables to plot the response curves of a model independently of the algorithm 
used for building the model. It therefore permits a direct comparison of predicted responses 
from the different statistical approaches on the same data.
}
\seealso{
\link[biomod2]{response.plot2}   
\link[biomod2]{BIOMOD_LoadModels}
}
