% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2plus.R
\name{varImportancePlot}
\alias{varImportancePlot}
\title{Plot variable importance from biomod2}
\usage{
varImportancePlot(biomodModelsOut, by = "all", sortVars = TRUE,
  filterAlgos = NULL, plotType = "facet_wrap", save = FALSE,
  plot = TRUE, outputFolder = getwd(), filename = "VarImportancePlot.png",
  ...)
}
\description{
Plot variable average importance scores across different PA sets, algorithms and 
evaluation rounds. \code{ggplot2} graphics are used to make the plots.
}
